# Vectorize normal_