--- a/aten/src/TH/generic/THVector.h
+++ b/aten/src/TH/generic/THVector.h
@@ -2,6 +2,8 @@
 #define TH_GENERIC_FILE "generic/THVector.h"
 #else
 
+struct THGenerator;
+
 TH_API void THVector_(fill)(real *x, const real c, const ptrdiff_t n);
 TH_API void THVector_(cadd)(real *z, const real *x, const real *y, const real c, const ptrdiff_t n);
 TH_API void THVector_(adds)(real *y, const real *x, const real c, const ptrdiff_t n);
@@ -11,6 +13,11 @@ TH_API void THVector_(cdiv)(real *z, const real *x, const real *y, const ptrdiff
 TH_API void THVector_(divs)(real *y, const real *x, const real c, const ptrdiff_t n);
 TH_API void THVector_(copy)(real *y, const real *x, const ptrdiff_t n);
 TH_API void THVector_(neg)(real *y, const real *x, const ptrdiff_t n);
+TH_API void THVector_(normal_fill)(real *data,
+                                   const int64_t size,
+                                   struct THGenerator *generator,
+                                   const real mean,
+                                   const real stddev);
 
 #if defined(TH_REAL_IS_SHORT) || defined(TH_REAL_IS_INT) || defined(TH_REAL_IS_LONG)
 TH_API void THVector_(abs)(real *y, const real *x, const ptrdiff_t n);