--- a/aten/src/TH/vector/AVX2.h
+++ b/aten/src/TH/vector/AVX2.h
@@ -7,10 +7,10 @@ struct THGenerator;
 
 void THDoubleVector_cadd_AVX2(double *z, const double *x, const double *y, const double c, const ptrdiff_t n);
 void THFloatVector_cadd_AVX2(float *z, const float *x, const float *y, const float c, const ptrdiff_t n);
-void THFloatTensor_normal_fill_AVX2(float *data,
-                                    const int size,
+void THFloatVector_normal_fill_AVX2(float *data,
+                                    const int64_t size,
                                     struct THGenerator *generator,
                                     const float mean,
-                                    const float stdv);
+                                    const float stddev);
 
 #endif