--- a/test/common.py
+++ b/test/common.py
@@ -21,8 +21,9 @@ from torch._six import string_classes
 
 torch.set_default_tensor_type('torch.DoubleTensor')
 
+
 parser = argparse.ArgumentParser(add_help=False)
-parser.add_argument('--seed', type=int, default=123)
+parser.add_argument('--seed', type=int, default=1234)
 parser.add_argument('--accept', action='store_true')
 args, remaining = parser.parse_known_args()
 SEED = args.seed