… aten/src/TH/vector/avx_mathfun.h