PyTorch is known to have high start-up times, to a large part due to slow tensor initialization/filling when calling `normal_`. [Torch's `normal` function](https://github.com/pytorch/pytorch/blob/master/aten/src/TH/THRandom.c#L253) uses the Box-Mueller transform to produce Gaussian distributed floats from uniform floats. I did some benchmarks and found that generating normal numbers took around 5 times longer than generating only uniform floats, suggesting that the current normal sampling code was the bottleneck.

This PR addresses this by introducing a vectorized version of the Box-Mueller transform that essentially does the same thing, but for 8 values at a time. This version is called only for floats, only if we have AVX2, only if there are more than 16 values (due to implementation) and only if the tensor is contiguous. However, this should cover like 90%-95% of real-world cases where it's currently slow.

My initial, small-scale benchmarks show a 5x-6x speed-up:

Before:
```
In [1]: import torch
In [2]: x = torch.Tensor(10000, 10000)
In [3]: %time _ = x.normal_(0, 1)
CPU times: user 3.45 s, sys: 111 ms, total: 3.57 s
Wall time: 3.57 s
```

After:
```
In [1]: import torch
In [2]: x = torch.Tensor(10000, 10000)
In [3]: %time _ = x.normal_(0, 1)
CPU times: user 611 ms, sys: 1.07 ms, total: 612 ms
Wall time: 613 ms
```

Which looks pretty good. I will see how this affects loading a model like imagenet or similar.

CC @zdevito