@goldsborough 2018-01-01
## Changes

Latest 2 commits make the following changes:

- Actually use the mean and standard deviation (was not using the variables before and thus generating unit Gaussian samples),
- Added a scalar `normal_fill` function that interleaves values just like the vectorized code, so that there is no difference in generated samples between AVX and non-AVX platforms. This function is also around 1.5x faster than the old version. Thus also non-AVX versions get a speedup (for contiguous tensors with at least 16 values).
- Implemented all the vector dispatch stuff
- Using `int64_t` for size
- Using `THAssert` instead of `assert`
- AVX code uses `_m256_loadu_ps` instead of `_m256_load_ps`, i.e. misaligned loads so that it works also for misaligned data,
- Using explicit FMA `_m256_fmadd_ps` instruction instead of multiply + add (clang doesn't do this automatically, GCC does ... it's less code anyway. Also FMA is available wherever AVX is on Intel and AMD chips, so should be fine). Marginally faster.

## Benchmarks

This time compiling with GCC 7.

### Microbenchmarks

10,000 x 10,000 (float/AVX): 3.3s -> 0.48s (6.875x speedup)
1,000 x 1,000 (float/AVX): 35ms -> 4.9ms (7.1x speedup)
10,000 x 10,000 (double/scalar): 3.2s -> 2.1s (1.5x speedup)

float/AVX here means it's using the vectorized version, double/scalar means it's the interleaved scalar function that I added, since the vectorized version is only called for floats.

### Imagenet Startup Times

VGG19: 5.61s -> 1.62s (3.5x speedup)
ResNet101: 2.4s -> 0.65s (3.7x speedup)
ResNet50: 1.21s -> 0.46s (2.6x speedup)

Happy new year 🎆 :tada: