@soumith 2018-01-01
@pytorchbot add to whitelist