@apaszke 2018-01-02
Should be ready to merge but the builds are failing now (probably because the seed we used previously is unlucky after this change)