@zdevito 2018-01-02
@pytorchbot retest this please