@goldsborough 2018-01-03
That worked, thanks Adam! Now one of the builds got stuck after a segfault in the dataloader tests. Seems flaky too.