@apaszke 2018-01-03
There's already other PR open that fixes the data loader thing. The test had a race condition